Or asked for this merge (TG mts0s5a2cdhz) so the fix can be tagged upstream.

THE BUG, measured live on the UTS 28 rebalance bot at 08:45 on 2026-09-08 and
for ~24 h before that:

  GetPosition -> json: cannot unmarshal number 50.0
  into Go struct field Position.result.leverage of type int

Deribit sends `"leverage":50.0`. The field was declared `int`, so encoding/json
rejected THE WHOLE RESPONSE — not just that field. GetAssetPositionSizeAndMarketValue
failed, the imbalance check errored every 5 minutes, and the bot placed ZERO orders
while reporting 1/1 Running with no restarts. Healthy and trading nothing.

Fixes both copies of the struct — deribit/api/position.go and deribit/ws/position.go
— because the REST and WS paths carry the same field and fixing one would leave the
other rejecting responses in exactly the same way.

⚠️ THIS MERGE ALONE DOES NOT FIX ANY BOT. go-deribit-rebalance requires
`bitbucket.org/ohm89/go-deribit v1.3.2` by module path and has no vendor dir and no
replace directive, so it resolves upstream regardless of what this repo contains.
The fix reaches production only when this commit is on the UPSTREAM main
(bitbucket.org/ohm89 or github.com/ohm89) and tagged, and gdrb's go.mod is bumped
to that tag. v1.3.3 does NOT contain it: that tag points at 58b681d (2026-04-16,
'changelog: add 2.0.0 entry'), which still declares Leverage as int on all three
remotes.

+++ deribit/api/position.go
@@ -38,7 +38,12 @@ type Position struct {
 	InstrumentName            string  `json:"instrument_name"`
 	InterestValue             float64 `json:"interest_value"`
 	Kind                      string  `json:"kind"`
-	Leverage                  int     `json:"leverage"`
+	// 🔴 float64, not int. Deribit sends `leverage` as a JSON number that can
+	// carry a fraction (observed live: 50.0). With `int`, encoding/json
+	// rejects the ENTIRE response — "cannot unmarshal number 50.0 into Go
+	// struct field Position.result.leverage of type int" — so GetPosition
+	// fails and every caller fails with it.
+	Leverage                  float64 `json:"leverage"`
 	MaintenanceMargin         float64 `json:"maintenance_margin"`
 	MarkPrice                 float64 `json:"mark_price"`
 	OpenOrdersMargin          float64 `json:"open_orders_margin"`
